requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
fonttools==4.39.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
neo4j==5.7.0
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.15.0
pymongo==4.3.3
pyodbc==4.0.35
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.9
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
